Corrected Degradation Rates
Corrected the degradation rates for this file. Before I took the median of all the genes instead of the median of the 203 TFs listed by Harbison. With that correction, I updated the calculations. In addition to these correction, I added a sheet at the end with the calculated degradation rates, rounded to four decimal places.